back/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
pylint
pytest
langchain_openai
langchain_core
langgraph
grandalf
google_auth_oauthlib
# flask_cors
cryptography
google-api-python-client
python_dotenv
requests
duckduckgo-search
# flask-cors
# flask-lambda
# aws-wsgi